consumer/requirements.txt
kafka_python==2.0.2
psycopg2_binary==2.9.9
